refactor: 💡 type move to contact type folder
