+++ src/pages/contact/type/index.ts
@@ -1,0 +1,9 @@
+export type InitialInput = {
+  name: string;
+  katakana: string;
+  mail: string;
+  phone: string;
+  period: string;
+  material: string;
+  postContent: string;
+};
